Fix absolute paths to relative (allow true static output)
Many of the paths were absolute, meaning that the generated site
couldn't be used in a true static setting - on disk, or located
at other than the root of the server. No functionality is lost,
just that relative URLS are used (unless canonifyURLS is set
in site config).

This is two changes - stop using .Site.BaseURL to form URLS by
hand, and use .RelPermalink instead of .Permalink.